+++ views/admin/en/module_options.php
@@ -9,16 +9,17 @@ $sLangName = "English";
 
 $aLang = [
     "charset"                                    => "UTF-8",
+    'SHOP_MODULE_GROUP_usercentrics_main'        => 'Usercentrics Integration',
     'SHOP_MODULE_GROUP_usercentrics_advanced'    => 'Advanced settings',
     'SHOP_MODULE_smartDataProtectorActive'       => 'Activate Usercentrics Smart Data Protector',
-    'HELP_SHOP_MODULE_smartDataProtectorActive'  => 'Disabling this function could have legal consequences',
+    'HELP_SHOP_MODULE_smartDataProtectorActive'  => ' Disabling this function could have legal consequences.',
     'SHOP_MODULE_usercentricsId'                 => 'Usercentrics Script ID',
     'SHOP_MODULE_usercentricsMode'               => 'Usercentrics Mode and Version',
     'SHOP_MODULE_usercentricsMode_CmpV1'         => 'CMP Version 1',
     'SHOP_MODULE_usercentricsMode_CmpV2'         => 'CMP Version 2',
     'SHOP_MODULE_usercentricsMode_CmpV2Legacy'   => 'CMP V2 for old Browsers',
-    'SHOP_MODULE_usercentricsMode_CmpTcf'        => 'TCF 2.0',
-    'SHOP_MODULE_usercentricsMode_CmpTcfLegacy'  => 'TCF 2.0 for old Browsers',
+    'SHOP_MODULE_usercentricsMode_CmpV2Tcf'      => 'TCF 2.0',
+    'SHOP_MODULE_usercentricsMode_CmpV2TcfLegacy' => 'TCF 2.0 for old Browsers',
     'SHOP_MODULE_usercentricsMode_Custom'        => 'Include usercentrics script by yourself',
-    'HELP_SHOP_MODULE_usercentricsMode'          => 'Choose the mode for Usercentrics. A documentation about versions, legacy browser support and TCF support can be found at <a href="https://docs.usercentrics.com/">docs.usercentrics.com</a>. Choose "Include usercentrics script by yourself" if the module itself should not add the usercentrics script.',
+    'HELP_SHOP_MODULE_usercentricsMode'          => 'Please choose the mode for Usercentrics! A documentation about versions, legacy browser support and TCF support can be found at <a target="_blank" href="https://docs.usercentrics.com/">docs.usercentrics.com</a>. Choose "Include usercentrics script by yourself" if the module itself should not add the usercentrics script.',
 ];